Factor out ADT, as is now Algae
