+++ test/functor_test.exs
@@ -1,11 +1,11 @@
-defmodule FunctorTest do
-  use ExUnit.Case
+# defmodule FunctorTest do
+#   use ExUnit.Case
 
-  import Witchcraft.Functor
-  import Witchcraft.Functor.Functions
-  import Witchcraft.Functor.Properties
+#   import Witchcraft.Functor
+#   import Witchcraft.Functor.Functions
+#   import Witchcraft.Functor.Properties
 
-  doctest Witchcraft.Functor
-  doctest Witchcraft.Functor.Functions
-  doctest Witchcraft.Functor.Properties
-end
+#   doctest Witchcraft.Functor
+#   doctest Witchcraft.Functor.Functions
+#   doctest Witchcraft.Functor.Properties
+# end
